requirements.txt
altair==5.5.0
altgraph==0.17.4
asgiref==3.7.2
attrs==25.3.0
Babel==2.14.0
blinker==1.9.0
cachetools==6.1.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.2.1
colorama==0.4.6
Django==5.0.1
django-bootstrap-modal-forms==3.0.4
django-cors-headers==4.3.1
django-filter==23.5
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.10
et-xmlfile==1.1.0
gitdb==4.0.12
GitPython==3.1.45
idna==3.6
inflection==0.5.1
Jinja2==3.1.6
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
MarkupSafe==3.0.2
mysql-connector-python==8.4.0
narwhals==2.0.1
numpy==1.26.3
openpyxl==3.1.2
packaging==24.0
pandas==2.2.2
pefile==2023.2.7
pillow==10.2.0
protobuf==6.31.1
pyarrow==21.0.0
pydeck==0.9.1
pyinstaller==6.5.0
pyinstaller-hooks-contrib==2024.3
PyJWT==2.8.0
pyserial==3.5
python-dateutil==2.9.0.post0
pytz==2023.3.post1
pywin32-ctypes==0.2.2
PyYAML==6.0.2
qrcode==8.2
referencing==0.36.2
reportlab==4.4.3
requests==2.31.0
rpds-py==0.27.0
setuptools==69.2.0
six==1.16.0
smmap==5.0.2
sqlparse==0.4.4
streamlit==1.48.0
tenacity==9.1.2
tkcalendar==1.6.1
toml==0.10.2
tornado==6.5.1
typing_extensions==4.14.1
tzdata==2023.4
uritemplate==4.1.1
urllib3==2.2.1
watchdog==6.0.0
whitenoise==6.6.0
xlrd==2.0.1
